Implement context manager API
Instead of passing around the token and re-establish the connection all
the time, we now pass around a session (which is a httpx Client).
The httpx library will reuse connections made to the server, improving
API call times.

The library can now be used as a context manager, meaning underlying
httpx connections will be terminated automatically.

In case the library is used as before, `.close()` or `.async_close()`
needs to be called to cleanup the connections.